requirements.txt
annotated-types==0.7.0
anthropic==0.83.0
anyio==4.12.1
asgiref==3.8.1
certifi==2026.1.4
chardet==7.0.1
charset-normalizer==3.4.5
distro==1.9.0
dj-database-url==2.1.0
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docstring_parser==0.17.0
et_xmlfile==2.0.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
openpyxl==3.1.2
packaging==26.0
pillow==11.2.1
psycopg2-binary==2.9.10
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
reportlab==4.0.9
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
whitenoise==6.6.0
